+++ plugins/auditoria-5s/src/components/KpiCard.tsx
@@ -1,0 +1,3 @@
+import { createAnalyticsKpiCard } from "@delpi/plugin-ui";
+
+export const KpiCard = createAnalyticsKpiCard("a5s");